+++ commands/General/invite.js
@@ -12,7 +12,7 @@ class Invite extends Command {
             dirname: __dirname,
             enabled: true,
             guildOnly: false,
-            aliases: [ "i", "add" ],
+            aliases: [ "i", "add", "vote" ],
             memberPermissions: [],
             botPermissions: [ "SEND_MESSAGES", "EMBED_LINKS" ],
             nsfw: false,